// backend/matching-websocket-service/src/dto/request.dto.ts
export type MatchRequest = {
    userId: string;
    topic: QuestionTopic;
    difficulty: QuestionComplexity;
    timestamp: number;
};

export type MatchRequestResponse = {
    message: string;
    error?: string;
    expiry?: number;
};

export type MatchFoundResponse = {
    matchedWithUserId: string;
    matchedTopic: string;
    matchedRoom: string;
}

export enum QuestionTopic {
    ANY = 'any',
    ARRAY = "array",
    BINARY = "binary",
    BINARY_SEARCH = "binary_search",
    BINARY_SEARCH_TREE = "binary_search_tree",
    BINARY_TREE = "binary_tree",
    DYNAMIC_PROGRAMMING = "dynamic_programming",
    GRAPH = "graph",
    GREEDY = "greedy",
    HASH_TABLE = "hash_table",
    HEAP = "heap",
    LINKED_LIST = "linked_list",
    MATH = "math",
    MATRIX = "matrix",
    QUEUE = "queue",
    RECURSION = "recursion",
    SORTING = "sorting",
    STACK = "stack",
    STRING = "string",
    TRIE = "trie",
}

export enum QuestionComplexity {
    ANY = 'any',
    EASY = "easy",
    MEDIUM = "medium",
    HARD = "hard",
}